chore: fixes for bad migrations
