+++ packages/client/src/components/engine/index.ts
@@ -1,3 +1,6 @@
 export * from "./engine-access-button";
 export * from "./engine-grid-item";
 export * from "./engine-header";
+export * from "./engine-migration-state-badge";
+export * from "./engine-migrations-summary";
+export * from "./engine-new-migration-dialog";
