+++ packages/client/src/pages/engine/engine.constants.ts
@@ -6,7 +6,7 @@ import { EngineFileManagerPage } from "./engine-file-manager-page";
 import { EngineFilePage } from "./engine-file-page";
 import { EngineMcpUsagePage } from "./engine-mcp-usage-page";
 import { EngineMetadataPage } from "./engine-metadata-page";
-import { EngineMigrationsPage } from "./engine-migrations-page";
+import { EngineMigrationsPage } from "./engine-migrations.page";
 import { EngineModelChatPage } from "./engine-model-chat-page";
 import { EngineOverviewPage } from "./engine-overview-page";
 import { EngineQAPage } from "./engine-qa-page";
